requirements.txt
streamlit==1.27.0
requests
Pillow
google-generativeai
python-dotenv